Refactor App state into hooks and settings dialogs
